+++ examples/autoscaling-group-windows/providers.tf
@@ -1,0 +1,11 @@
+provider "aws" {
+  version    = "~> 2"
+  region     = "us-east-2"
+  access_key = var.access_key
+  secret_key = var.secret_key
+
+  assume_role {
+    role_arn     = "arn:aws:iam::700633540182:role/OrganizationAccountAccessRole"
+    session_name = "FXTestSandbox"
+  }
+}