+++ CHANGELOG.md
@@ -5,6 +5,7 @@ fix: fix AutoScaling group creation with latest AWS provider
 fix: fix `cpu_options` for ASG
 fix: fix not working dynamic blocks throughout the module
 tech: update possible version for AWS provider to 2.60 and up
+test: rename `no-instances-no-volumes` example to `disabled`
 
 6.1.0
 =====